d500: inline sensors_config_mode_option; PID-gate registration; reorder UI
Address human review feedback on PR #15499:

- Move sensors_config_mode_option class into d500-options.{h,cpp} alongside
  the other D500 option classes (temperature_xu_option et al.). Deletes the
  standalone d500-sensors-config-mode.{h,cpp} pair.
- Register from d500_device::init() gated by a new d500_sensors_config_mode_pids
  set in d500-private.h (D535/D585 2C/3C/F + protos), mirroring the
  d500_projector_temperature_pids pattern already used for XU 0x16.
  The FW capability probe (try/query() before register_option) stays as a
  safety net for old FW builds on SKUs in that PID set.
- Drop the three per-SKU register_sensors_config_mode_option() calls from
  d500-factory.cpp; remove the header include; drop the file entries from
  d500/CMakeLists.txt.
- Reorder drawing_order in device-model.cpp so SENSORS_CONFIG_MODE renders
  first in the Stereo Module pane (above VISUAL_PRESET / EMITTER_ENABLED /
  ENABLE_AUTO_EXPOSURE / DEPTH_AUTO_EXPOSURE_MODE).
