+++ src/ds/d500/d500-private.h
@@ -62,6 +62,21 @@ namespace librealsense
             D585_3C_PROTO_PID
         };
 
+        // d500 PIDs that expose the DUAL_RGB_MODE XU (0x12) selector — 2C/3C variants that can
+        // switch between the dedicated-color-sensor (3C) and dual-RGB (2C) FW configurations.
+        // The runtime FW query in sensors_config_mode_option::query() still gates final
+        // registration, so this list only decides which SKUs are worth probing at all.
+        static const std::set<std::uint16_t> d500_sensors_config_mode_pids = {
+            D535_2C_PID,
+            D535_3C_PID,
+            D535F_PID,
+            D585_2C_PID,
+            D585_3C_PID,
+            D585F_PID,
+            D585_2C_PROTO_PID,
+            D585_3C_PROTO_PID
+        };
+
         // d500 PIDs that expose the projector temperature via HKR selector 0x16
         static const std::set<std::uint16_t> d500_projector_temperature_pids = {
             D585S_PID,
